+++ src/@locale/components/index.js
@@ -1,17 +1,21 @@
 import buttons    from './buttons'
 import dropdowns  from './dropdowns'
+import input      from './input'
 import modal      from './modal'
 import pagination from './pagination'
 import picker     from './picker'
+import sortable   from './sortable'
 import splitPanel from './splitPanel'
 
 const components =
 {
     buttons ,
     dropdowns,
+    input ,
     modal ,
     pagination ,
     picker ,
+    sortable ,
     splitPanel ,
 };